==> scripts/0_extract_reddit.ipynb <==
# 0. Script for extracting Reddit data

## Set Up

import numpy as np
import pandas as pd
import requests
from time import sleep
import datetime

## Extracting submissions (posts)

To standardise the content that users will be responding to in their comments, I will be extracting posts that have the **same content on both Reddit and Facebook**. These posts must meet the following conditions:
- Contain only **links** to Straits Times and Channel News Asia articles reporting on official Ministry of Health (MOH) announcements of COVID-19 restrictions in Singapore. There should be no further discussion about the article in the post.
- The linked articles should only be about **imposition/tightening** of restrictions and not loosening of the restrictions, and should only report the announcement of the restrictions i.e. no opinion pieces or follow-up comments by ministers).

Querying the Pushshift Reddit API, I searched for posts from May 2021 to May 2022 that contained at least one of the following search terms in the title: 'covid', 'phase' (as Singapore's restrictions were announced as part of Phases), 'measures', 'restrictions', 'suspended' (some restrictions involved suspension of visits to medical institutions).

res_dfs = []

date = datetime.datetime(2021,5,1,0,0,0)
while(date <= datetime.datetime(2022,5,1,0,0,0)):
    start_date = date
    date += datetime.timedelta(days=5)
    end_date = date 

    url = 'https://api.pushshift.io/reddit/search/submission/'
    query = {'subreddit': 'singapore',
             'after': start_date,
             'before': end_date,
             'title': 'covid',
             'num_comments': '>0',
             'fields': ['id', 'author', 'title', 'num_comments', 'score', 'url', 'full_link', 'created_utc'],
             'size': 100,
             'sort': 'asc',
             'sort_type': 'created_utc'}
    res = requests.get(url=url, params=query)
    res = res.json()
    res_df = pd.DataFrame(res['data'])
    res_dfs.append(res_df)
    
    sleep(10)
    
res_dfs = pd.concat(res_dfs)
res_dfs.reset_index(drop=True, inplace=True)

res_selected = res_dfs[res_dfs['num_comments'] > 0]
res_selected = res_selected[res_selected['url'].str.contains('channelnewsasia|straitstimes|cna', na = False)]

res_selected.to_csv('../data/posts_reddit_covid_all.csv')

## Extracting comments

After extracting all the Reddit posts with the designated search terms, I manually looked through all of the post titles and selected only those which met the aforementioned criteria - posts that only contained links to articles reporting on official MOH announcements of tightening of restrictions.

I then found links to the corresponding Facebook posts from official Straits Times and Channel News Asia Facebook pages that contained the same article as the Reddit posts. The reddit post data and the corresponding Facebook post links for each Reddit post were saved as `posts_reddit_fb_selected.csv`.

### Import selected reddit posts data

reddit_selected = pd.read_csv('../data/posts_reddit_fb_selected.csv')

reddit_selected

### Get comment IDs for each post

Having obtained the post ID for each post, I then queried the Pushshift API to obtain all the comment IDs across all selected posts, so that I could later use the comment ID to obtain the comment text.

comment_ids = {'post_id': [], 'comment_id': []}

for i in reddit_selected['id']:
    url = 'https://api.pushshift.io/reddit/submission/comment_ids/' + i
    res = requests.get(url=url)
    res = res.json()
    for j in res['data']:
        comment_ids['post_id'].append(i)
        comment_ids['comment_id'].append(j)
        
comment_ids_df = pd.DataFrame(comment_ids)

comment_ids_df

comment_ids_df.to_csv('../data/comments_reddit_ids.csv')

### Get comments by comment ID

Using the extracted comment IDs, I queried the Pushshift API to obtain the corresponding comment text. Due to query limits, the 9000+ comments were extracted 1000 at a time, then concatenated into a single dataframe at the end.

comment_ids_df = pd.read_csv('../data/comments_reddit_ids.csv', index_col=0)
comment_ids_df.head()

comment_ids_df.shape

# Extract 1000 comments at a time
comments = {'comment_id': [], 'text': []}

idx = 8000
while (idx < 9692):
    start = idx
    idx += 1
    end = idx
    sample = comment_ids_df['comment_id'].iloc[start:end]
    
    for c_id in sample:
        url = 'https://api.pushshift.io/reddit/search/comment/'
        query = {'ids': c_id}
        res = requests.get(url=url, params=query)
        res = res.json()
        
        try:
            text = res['data'][0]['body']
        except IndexError:
            text = ''
        
        comments['comment_id'].append(str(c_id))
        comments['text'].append(str(text))
        comments_df_9 = pd.DataFrame(comments)
    
    sleep(2)
    
comments_df_9.to_csv('../data/reddit_comments_9.csv', encoding='utf-8-sig')
display(comments_df_9)

### Concatenating all comments

comments_df_1 = pd.read_csv('../data/comments_reddit_1.csv', index_col=0)
comments_df_2 = pd.read_csv('../data/comments_reddit_2.csv', index_col=0)
comments_df_3 = pd.read_csv('../data/comments_reddit_3.csv', index_col=0)
comments_df_4 = pd.read_csv('../data/comments_reddit_4.csv', index_col=0)
comments_df_5 = pd.read_csv('../data/comments_reddit_5.csv', index_col=0)
comments_df_6 = pd.read_csv('../data/comments_reddit_6.csv', index_col=0)
comments_df_7 = pd.read_csv('../data/comments_reddit_7.csv', index_col=0)
comments_df_8 = pd.read_csv('../data/comments_reddit_8.csv', index_col=0)
comments_df_9 = pd.read_csv('../data/comments_reddit_9.csv', index_col=0)

comments_df_list = [comments_df_1, comments_df_2, comments_df_3, comments_df_4, comments_df_5, comments_df_6, comments_df_7,
                   comments_df_8, comments_df_9]
comments_dfs = pd.concat(comments_df_list, axis=0, ignore_index=True)

# Remove empty, deleted or removed comments
comments_dfs = comments_dfs.drop(comments_dfs[(comments_dfs['text']=='[removed]')|(comments_dfs['text']=='[deleted]')|(comments_dfs['text'].isna())].index)

# Remove bot comments that repost the article text 
comments_dfs = comments_dfs[comments_dfs['text'].str.contains('https://github.com/fterh/sneakpeek')==False]

comments_dfs.shape

comments_dfs.to_csv('../data/comments_reddit_all.csv', encoding='utf-8-sig')
