==> scripts/2_word_frequency.ipynb <==
# 2. Script for obtaining word frequencies in comment corpora

## Set Up

import pandas as pd

path = "C:\\Users\\nicol\\Documents\\2021-2022\\QMSS\\Spring'22\\QMSS 5999 Thesis\\data\\"

# Import comment data
reddit_raw = pd.read_csv(path+'comments_reddit_all.csv', index_col=0, dtype={'comment_id': 'str', 'text': 'str'})
fb_raw = pd.read_csv(path+'comments_fb_all.csv', index_col=0, dtype={'comment_id': 'str', 'text': 'str'})
fb_raw = fb_raw[['comment_id', 'text']]

# Concatenate both FB and Reddit data into one DataFrame
all_raw = pd.concat([reddit_raw, fb_raw], ignore_index=True)

# Inspect DataFrame and get dimensions
display(all_raw.head())
display(all_raw.tail())
display(all_raw.shape)

## Preprocessing text

### Clean text

# Function for cleaning text
def clean_text_freq(text):
    import re
    import html
    text1 = html.unescape(text) # remove HTML encoded characters
    text2 = re.sub(r'â€™', '\'', text1) # replce â€™ with apostrophe
    text3 = re.sub(r'\n', '', text2) # remove excess line breaks between links
    text4 = re.sub(r'\n', ' ', text3) # replace remaining line breaks with spaces
    text5 = re.sub(r'#\S+', '', text4) # remove hashtags
    text6 = re.sub(r'http\S+', '', text5) # remove links starting with http
    text7 = re.sub(r'\S+\.com\S+', '', text6) # remove links not starting with http
    text8 = re.sub(r'[^A-Za-z]+', ' ', text7).strip().lower() # remove punctuations and digits 
    return(text8)

# Apply function to DataFrame
all_raw['text_clean_freq'] = all_raw.text.apply(clean_text_freq)

# Inspect DataFrame
all_raw.head()

### Remove stop words

# Function for removing stop words
def remove_sw(text):
    from nltk.corpus import stopwords
    sw = set(stopwords.words('English'))
    text_filtered = [word for word in text.split() if not word in sw] 
    text_filtered = ' '.join(text_filtered) 
    return(text_filtered)

# Apply function to DataFrame
all_raw['text_no_sw'] = all_raw.text_clean_freq.apply(remove_sw)

# Inspect DataFrame
all_raw.head()

### Tokenize comment corpus

# Transform text in DataFrame to a single string
all_raw_string = ' '.join(all_raw['text_no_sw']) 

# Tokenise the whole corpus
from nltk.tokenize import word_tokenize
tokens_all = word_tokenize(all_raw_string)
print(len(tokens_all))

# Obtain tokens not in NLTK's English dictionary (to identify non-English words that need to be parsed/replaced)
import nltk
from nltk.corpus import words
from nltk.corpus import wordnet

nltk_words = words.words()
tokens_no_eng = [tok for tok in tokens_all if not tok in nltk_words if not wordnet.synsets(tok)]

## Obtain top 1000 tokens by frequency

### Whole corpus

# Obtain top 1000 tokens from the whole corpus
from nltk.probability import FreqDist
fdist_all = FreqDist(tokens_all)
tokens_all_top1000 = fdist_all.most_common(1000)

# Inspect top 20 
tokens_all_top1000[0:20]

# Inspect bottom 20
tokens_all_top1000[-20:]

# Convert to DataFrame and inspect
tokens_all_top1000_df = pd.DataFrame(tokens_all_top1000, columns=['token', 'freq'])
tokens_all_top1000_df.head()

# Export as CSV
tokens_all_top1000_df.to_csv(path+'comments_all_tokens_top1000.csv', encoding='utf-8-sig')

### Non-English corpus

# Obtain top 1000 non-English tokens
from nltk.probability import FreqDist
fdist_no_eng = FreqDist(tokens_no_eng)
tokens_no_eng_top1000 = fdist_no_eng.most_common(1000)

# Inspect top 20 
tokens_no_eng_top1000[0:20]

These non-English tokens will guide me in constructing a dictionary to replace these words with their meaning, to allow the sentiment analysis algorithms to better parse the text and to improve the results. 

Additionally, this list helps me to decide which tokens should be kept:
- For instance, 'cb' is a top token but it has multiple meanings in the corpora as it refers to either a swear word ('cb' short for 'chee bye' which is equivalent in vulgarity to the F word), or the Circuit Breaker lockdown restrictions in April 2020 (which the government and citizens have commonly abbreviated into 'CB'). Given the confusion, I will eliminate this specifc token from the list. I will only retain 'cb' if it is part of another token like 'ccb' which definitely refers to the swear phrase ('chao chee bye') and not the restrictions. 
- I will also not replace tokens which do not have a concise explanation/replacement in the English language, or tokens that are proper nouns.
- Internet slang abbreviations e.g. lol, lmao, etc. will not be replaced as they can be captured by VADER.

# Convert to DataFrame and export as CSV
tokens_no_eng_top1000_df = pd.DataFrame(tokens_no_eng_top1000, columns=['token', 'freq'])
tokens_no_eng_top1000_df.to_csv(path+'comments_all_tokens_no_eng_top1000.csv', encoding='utf-8-sig')
