==> scripts/3_sentiment_analysis.ipynb <==
# 3. Script for conducting sentiment analysis on comment data

## Set Up

import pandas as pd

pd.set_option('display.max_colwidth', None)

path = "C:\\Users\\nicol\\Documents\\2021-2022\\QMSS\\Spring'22\\QMSS 5999 Thesis\\data\\"

# Import comment data
reddit = pd.read_csv(path+'comments_reddit_all.csv', index_col=0, dtype={'comment_id': 'str', 'text': 'str'})
fb = pd.read_csv(path+'comments_fb_all.csv', index_col=0, dtype={'comment_id': 'str', 'text': 'str'})
fb = fb[['comment_id', 'text']]

display(reddit.head())
display(reddit.shape)
display(fb.head())
display(fb.shape)

## Preprocessing text

### Clean text

# Function for cleaning text
def clean_text_sa(text):
    import re
    import html
    text1 = html.unescape(text) # remove HTML encoded characters
    text2 = re.sub(r'â€™', '\'', text1) # replce â€™ with apostrophe
    text3 = re.sub(r'\n', ' ', text2) # replace line breaks with spaces
    text4 = re.sub(r'#\S+', '', text3) # remove hashtags
    text5 = re.sub(r'http\S+', '', text4) # remove links starting with http
    text6 = re.sub(r'\S+\.com\S+', '', text5) # remove links not starting with http
    return(text6)

# Apply function to DataFrames
reddit['text_clean_sa'] = reddit.text.apply(clean_text_sa)
fb['text_clean_sa'] = fb.text.apply(clean_text_sa)

### Replacing words/phrases according using self-built dictionary

# Import dictionary
dict_clean_df = pd.read_csv(path+'dict.csv', keep_default_na=False)
dict_clean = dict(dict_clean_df.values)

dict_clean

# Function for replacing text using dictionary mapping
def replace_text(text):
    import re
    for orig, new in dict_clean.items():
        text = re.sub(r'\b{}(\b|^A-Za-z0-9)'.format(orig.strip()), new, text)
    return(text)

# Apply function to FB and Reddit DataFrames
reddit['text_replaced'] = reddit.text_clean_sa.apply(replace_text)
fb['text_replaced'] = fb.text_clean_sa.apply(replace_text)

# Inspect DataFrames
display(reddit.head())
display(fb.head())

# Export cleaned text corpora to CSV
reddit.to_csv(path+'comments_reddit_all_clean.csv', header=True, index=True, encoding='utf-8-sig')
fb.to_csv(path+'comments_fb_all_clean.csv', header=True, index=True, encoding='utf-8-sig')

## Sentiment analysis 

### VADER

import pandas as pd

pd.set_option('display.max_colwidth', None)

path = "C:\\Users\\nicol\\Documents\\2021-2022\\QMSS\\Spring'22\\QMSS 5999 Thesis\\data\\"

# Import score data
reddit = pd.read_csv(path+'comments_reddit_all_score.csv', index_col=0, dtype={'comment_id': 'str', 'text': 'str'}, keep_default_na=False)
fb = pd.read_csv(path+'comments_fb_all_score.csv', index_col=0, dtype={'comment_id': 'str', 'text': 'str'}, keep_default_na=False)
reddit = reddit[['comment_id', 'text', 'text_clean_sa', 'text_replaced']]
fb = fb[['comment_id', 'text', 'text_clean_sa', 'text_replaced']]

# Define function for scoring sentiment using VADER
def score_vader(text):
    from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
    from decimal import Decimal
    analyzer = SentimentIntensityAnalyzer()
    scores = analyzer.polarity_scores(text)
    score = scores['compound']
    return(round(Decimal(score), 4))

# Apply function to DataFrames
reddit['score_vader'] = reddit.text_replaced.apply(score_vader)
fb['score_vader'] = fb.text_replaced.apply(score_vader)

# Inspect DataFrames
display(reddit.head())
display(fb.head())

### TextBlob

# Define function for scoring sentiment using TextBlob
def score_tb(text):
    from textblob import TextBlob
    from decimal import Decimal
    score = TextBlob(text).sentiment.polarity
    return(round(Decimal(score), 4))

# Apply function to DataFrames
reddit['score_tb'] = reddit.text_replaced.apply(score_tb)
fb['score_tb'] = fb.text_replaced.apply(score_tb)

# Inspect DataFrames
display(reddit.head())
display(fb.head())

# Export text corpora with scores to CSV
reddit.to_csv(path+'comments_reddit_all_score.csv', header=True, index=True, encoding='utf-8-sig')
fb.to_csv(path+'comments_fb_all_score.csv', header=True, index=True, encoding='utf-8-sig')
